Adopt wuserlibrarypath()
Replace calls to wusergnusteppath() which just append "/Library" with calls to wuserlibrarypath(). Alsö, replace "/WindowMaker" with "/"PACKAGE.

This accounts for the situation where the user's Library and Defaults directories may not share a recent common ancestor.
This continues to honor WMAKER_USER_ROOT by way of wuserlibrarypath()'s fallback to wusergnusteppath().