WINGs: restructure wmkdirhier() and wrmdirkier()
Use wuserdatapath() and wuserdefautspathfordomain("") to limit scope instead of wusergnusteppath(), and rewrite to recursively create directories instead of pointer arithmatic.

This accounts for the situation where the user's Library and Defaults directories may not share a recent common ancestor.
This continues to honor WMAKER_USER_ROOT by way of wuserdatapath()'s fallback to wusergnusteppath().